bafl-app-updater/requirements.txt
streamlit>=1.30.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
